Refactor print statements, improve error handling

+++ run.py
@@ -34,7 +34,7 @@ def main():
 
         if os.path.isfile(cached_file_path):
             shutil.copyfile(cached_file_path, dst_file_path)
-            console.print(f"[bold blue]{DATE} - Map Download - Finished from cache")
+            console.print(f"[bold green]{DATE} - Map Download - Finished from cache")
             continue
 
         with open(f"input_file.txt", "w") as f:
@@ -43,8 +43,9 @@ def main():
             )
 
         with console.status(
-            f"[bold green]{DATE} - Map Download - Uploading map request"
-        ) as status:
+            f"[bold blue]{DATE} - Map Download - Uploading request",
+            spinner_style="bold blue",
+        ):
             while True:
                 try:
                     result = subprocess.run(
@@ -56,21 +57,29 @@ def main():
                     assert result.returncode == 0
                     break
                 except:
-                    console.print_exception(show_locals=True)
-                    # error, when calling login.sh too fast in a row
-                    console.print(
-                        f"[bold yellow]{DATE} - Map Download - too frequent file "
-                        + "request, cooling down a minute"
-                    )
-                    time.sleep(60)
+                    if "Access failed: 553" in result.stderr.decode():
+                        console.print(
+                            f"[bold blue]{DATE} - Map Download - "
+                            + "Request was rate limited, waiting 1 minute"
+                        )
+                        time.sleep(60)
+                    else:
+                        console.print(
+                            f"[bold red]{DATE} - Map Download - "
+                            + f"Uploading request failed: {result.stderr}"
+                        )
+                        return
 
             os.remove("input_file.txt")
 
         map_file = f"maps_{lat_str}{lng_str}_{DATE}_{DATE}.tar"
 
-        with console.status(f"[bold green]{DATE} - Downloading map file") as status:
+        with console.status(
+            f"[bold blue]{DATE} - Map Download - Downloading file",
+            spinner_style="bold blue",
+        ):
             while True:
-                subprocess.run(
+                result = subprocess.run(
                     [
                         "wget",
                         "--user",
@@ -82,24 +91,42 @@ def main():
                     stdout=subprocess.PIPE,
                     stderr=subprocess.PIPE,
                 )
+                if result.returncode != 0:
+                    console.print(
+                        f"[bold red]{DATE} - Map Download - "
+                        + f"command 'wget ...' returned "
+                        + f"a non-zero exit code: {result.stderr}"
+                    )
+                    return
                 if os.path.isfile(map_file):
                     break
                 time.sleep(8)
 
         with console.status(
-            f"[bold green]{DATE} - Map Download - Processing map file"
-        ) as status:
-            p = subprocess.run(
+            f"[bold blue]{DATE} - Map Download - Processing file",
+            spinner_style="bold blue",
+        ):
+            result = subprocess.run(
                 ["tar", "-xvf", map_file],
                 stderr=subprocess.PIPE,
                 stdout=subprocess.PIPE,
             )
+            if result.returncode != 0:
+                console.print(
+                    f"[bold red]{DATE} - Map Download - "
+                    + f"command 'tar ...' returned "
+                    + f"a non-zero exit code: {result.stderr}"
+                )
+                return
+            assert os.path.isfile(f"L1{DATE}.map")
+
+            # store generated file in internal cache
             if not os.path.isfile(cached_file_path):
                 shutil.copyfile(f"L1{DATE}.map", cached_file_path)
             os.rename(f"L1{DATE}.map", dst_file_path)
             os.remove(map_file)
 
-        console.print(f"[bold blue]{DATE} - Map Download - Finished")
+        console.print(f"[bold green]{DATE} - Map Download - Finished from source")
 
 
 if __name__ == "__main__":
